salvo cambios

+++ stats/getItemDfs.py
@@ -11,17 +11,17 @@ df_items = pd.DataFrame(get_all_items_data(final_form_only=True)).T[[
     "name", "id", "price", 'final_form', "tags"]]
 # Items de vision
 trinkets = df_items.loc[((df_items['tags'].astype(
-    str).str.contains("Trinket")))]['id'].tolist()
+    str).str.contains("Trinket")))]['id'].astype(int).tolist()
 # Items finales
 final_form_items = df_items.loc[~(df_items['tags'].astype(
     str).str.contains("Trinket|Consumable")) &
-    ~(df_items['name'].astype(str).str.contains("Encantamiento"))]['id'].unique()
+    ~(df_items['name'].astype(str).str.contains("Encantamiento"))]['id'].astype(int).unique()
 # Botas
 boots = df_items.loc[(df_items['tags'].astype(
     str).str.contains("Boots"))]['id'].tolist()
 # GoldPer
 support_items = df_items.loc[(
-    (df_items['tags'].astype(str).str.contains("GoldPer")))]['id'].unique()
+    (df_items['tags'].astype(str).str.contains("GoldPer")))]['id'].astype(int).unique()
 # Jungle items
 jungle_items = df_items.loc[(df_items['tags'].astype(
-    str).str.contains("Jungle"))]['id'].unique()
+    str).str.contains("Jungle"))]['id'].astype(int).unique()
